Extract v2 parser ref helpers
